requirements_test.txt
pytest-homeassistant-custom-component==0.4.0
msmart==0.1.24
